# swagger-gen-fe/generated-code/model/contentDisposition.ts
/**
 * My API
 * This is a description of my API.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ContentDispositionCharset } from './contentDispositionCharset';


export interface ContentDisposition { 
    type?: string;
    name?: string;
    filename?: string;
    charset?: ContentDispositionCharset;
    size?: number;
    creationDate?: string;
    modificationDate?: string;
    readDate?: string;
    attachment?: boolean;
    inline?: boolean;
    formData?: boolean;
}