# swagger-gen-fe/generated-code/model/httpStatusCode.ts
/**
 * My API
 * This is a description of my API.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface HttpStatusCode { 
    error?: boolean;
    is1xxInformational?: boolean;
    is3xxRedirection?: boolean;
    is2xxSuccessful?: boolean;
    is5xxServerError?: boolean;
    is4xxClientError?: boolean;
}