# swagger-gen-fe/generated-code/model/mediaType.ts
/**
 * My API
 * This is a description of my API.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ContentDispositionCharset } from './contentDispositionCharset';


export interface MediaType { 
    type?: string;
    subtype?: string;
    parameters?: { [key: string]: string; };
    qualityValue?: number;
    subtypeSuffix?: string;
    wildcardType?: boolean;
    wildcardSubtype?: boolean;
    concrete?: boolean;
    charset?: ContentDispositionCharset;
}