# swagger-gen-fe/generated-code/model/responseEntityUser.ts
/**
 * My API
 * This is a description of my API.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { User } from './user';
import { HttpStatusCode } from './httpStatusCode';
import { ResponseEntityVoidHeaders } from './responseEntityVoidHeaders';


export interface ResponseEntityUser { 
    headers?: ResponseEntityVoidHeaders;
    body?: User;
    statusCode?: HttpStatusCode;
    statusCodeValue?: number;
}