# swagger-gen-fe/swagger/swagger-spec-generated.json
{
  "openapi" : "3.0.1",
  "info": {
    "title": "My API",
    "description": "This is a description of my API.",
    "version": "1.0.0"
  },
  "paths" : {
...
    "/users/{userid}" : {
      "get" : {
        "tags" : [ "Users" ],
        "summary" : "Get an user by ID",
        "operationId" : "getUser",
        "parameters" : [ {
          "name" : "userid",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResponseEntityUser"
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "ContentDisposition" : {
        "type" : "object",
        "properties" : {
          "type" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "filename" : {
            "type" : "string"
          },
          "charset" : {
            "type" : "object",
            "properties" : {
              "registered" : {
                "type" : "boolean"
              }
            }
          },
          "size" : {
            "type" : "integer",
            "format" : "int64"
          },
          "creationDate" : {
            "type" : "string",
            "format" : "date-time"
          },
          "modificationDate" : {
            "type" : "string",
            "format" : "date-time"
          },
          "readDate" : {
            "type" : "string",
            "format" : "date-time"
          },
          "attachment" : {
            "type" : "boolean"
          },
          "inline" : {
            "type" : "boolean"
          },
          "formData" : {
            "type" : "boolean"
          }
        }
      },
...
      "HttpMethod" : {
        "type" : "object"
      },
      "HttpRange" : {
        "type" : "object"
      },
      "HttpStatusCode" : {
        "type" : "object",
        "properties" : {
          "error" : {
            "type" : "boolean"
          },
          "is1xxInformational" : {
            "type" : "boolean"
          },
          "is3xxRedirection" : {
            "type" : "boolean"
          },
          "is2xxSuccessful" : {
            "type" : "boolean"
          },
          "is5xxServerError" : {
            "type" : "boolean"
          },
          "is4xxClientError" : {
            "type" : "boolean"
          }
        }
      },
      "MediaType" : {
        "type" : "object",
        "properties" : {
          "type" : {
            "type" : "string"
          },
          "subtype" : {
            "type" : "string"
          },
          "parameters" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "qualityValue" : {
            "type" : "number",
            "format" : "double"
          },
          "subtypeSuffix" : {
            "type" : "string"
          },
          "wildcardType" : {
            "type" : "boolean"
          },
          "wildcardSubtype" : {
            "type" : "boolean"
          },
          "concrete" : {
            "type" : "boolean"
          },
          "charset" : {
            "type" : "object",
            "properties" : {
              "registered" : {
                "type" : "boolean"
              }
            }
          }
        }
      },
...
      "User" : {
        "type" : "object",
        "properties" : {
          "username" : {
            "type" : "string"
          }
        }
      },
      "ResponseEntityUser" : {
        "type" : "object",
        "properties" : {
          "headers" : {
            "type" : "object",
            "properties" : {
              "empty" : {
                "type" : "boolean"
              },
              "location" : {
                "type" : "string",
                "format" : "uri"
              },
              "host" : {
                "type" : "object",
                "properties" : {
                  "address" : {
                    "type" : "object",
                    "properties" : {
                      "canonicalHostName" : {
                        "type" : "string"
                      },
                      "hostAddress" : {
                        "type" : "string"
                      },
                      "address" : {
                        "type" : "array",
                        "items" : {
                          "type" : "string",
                          "format" : "byte"
                        }
                      },
                      "hostName" : {
                        "type" : "string"
                      },
                      "linkLocalAddress" : {
                        "type" : "boolean"
                      },
                      "anyLocalAddress" : {
                        "type" : "boolean"
                      },
                      "multicastAddress" : {
                        "type" : "boolean"
                      },
                      "loopbackAddress" : {
                        "type" : "boolean"
                      },
                      "siteLocalAddress" : {
                        "type" : "boolean"
                      },
                      "mcglobal" : {
                        "type" : "boolean"
                      },
                      "mcnodeLocal" : {
                        "type" : "boolean"
                      },
                      "mclinkLocal" : {
                        "type" : "boolean"
                      },
                      "mcsiteLocal" : {
                        "type" : "boolean"
                      },
                      "mcorgLocal" : {
                        "type" : "boolean"
                      }
                    }
                  },
                  "port" : {
                    "type" : "integer",
                    "format" : "int32"
                  },
                  "unresolved" : {
                    "type" : "boolean"
                  },
                  "hostName" : {
                    "type" : "string"
                  },
                  "hostString" : {
                    "type" : "string"
                  }
                }
              },
              "all" : {
                "type" : "object",
                "additionalProperties" : {
                  "type" : "string"
                },
                "writeOnly" : true
              },
              "lastModified" : {
                "type" : "integer",
                "format" : "int64"
              },
              "date" : {
                "type" : "integer",
                "format" : "int64"
              },
              "contentLength" : {
                "type" : "integer",
                "format" : "int64"
              },
              "accessControlAllowCredentials" : {
                "type" : "boolean"
              },
              "accessControlAllowHeaders" : {
                "type" : "array",
                "items" : {
                  "type" : "string"
                }
              },
              "accessControlAllowMethods" : {
                "type" : "array",
                "items" : {
                  "$ref" : "#/components/schemas/HttpMethod"
                }
              },
              "accessControlAllowOrigin" : {
                "type" : "string"
              },
              "accessControlExposeHeaders" : {
                "type" : "array",
                "items" : {
                  "type" : "string"
                }
              },
              "accessControlRequestHeaders" : {
                "type" : "array",
                "items" : {
                  "type" : "string"
                }
              },
              "accessControlRequestMethod" : {
                "$ref" : "#/components/schemas/HttpMethod"
              },
              "connection" : {
                "type" : "array",
                "items" : {
                  "type" : "string"
                }
              },
              "acceptLanguageAsLocales" : {
                "type" : "array",
                "items" : {
                  "type" : "object",
                  "properties" : {
                    "language" : {
                      "type" : "string"
                    },
                    "displayName" : {
                      "type" : "string"
                    },
                    "country" : {
                      "type" : "string"
                    },
                    "variant" : {
                      "type" : "string"
                    },
                    "script" : {
                      "type" : "string"
                    },
                    "unicodeLocaleAttributes" : {
                      "uniqueItems" : true,
                      "type" : "array",
                      "items" : {
                        "type" : "string"
                      }
                    },
                    "unicodeLocaleKeys" : {
                      "uniqueItems" : true,
                      "type" : "array",
                      "items" : {
                        "type" : "string"
                      }
                    },
                    "displayLanguage" : {
                      "type" : "string"
                    },
                    "displayScript" : {
                      "type" : "string"
                    },
                    "displayCountry" : {
                      "type" : "string"
                    },
                    "displayVariant" : {
                      "type" : "string"
                    },
                    "extensionKeys" : {
                      "uniqueItems" : true,
                      "type" : "array",
                      "items" : {
                        "type" : "string"
                      }
                    },
                    "iso3Language" : {
                      "type" : "string"
                    },
                    "iso3Country" : {
                      "type" : "string"
                    }
                  }
                }
              },
              "acceptPatch" : {
                "type" : "array",
                "items" : {
                  "$ref" : "#/components/schemas/MediaType"
                }
              },
              "accessControlMaxAge" : {
                "type" : "integer",
                "format" : "int64"
              },
              "acceptCharset" : {
                "type" : "array",
                "items" : {
                  "type" : "object",
                  "properties" : {
                    "registered" : {
                      "type" : "boolean"
                    }
                  }
                }
              },
              "bearerAuth" : {
                "type" : "string",
                "writeOnly" : true
              },
              "cacheControl" : {
                "type" : "string"
              },
              "contentDisposition" : {
                "$ref" : "#/components/schemas/ContentDisposition"
              },
              "contentLanguage" : {
                "type" : "object",
                "properties" : {
                  "language" : {
                    "type" : "string"
                  },
                  "displayName" : {
                    "type" : "string"
                  },
                  "country" : {
                    "type" : "string"
                  },
                  "variant" : {
                    "type" : "string"
                  },
                  "script" : {
                    "type" : "string"
                  },
                  "unicodeLocaleAttributes" : {
                    "uniqueItems" : true,
                    "type" : "array",
                    "items" : {
                      "type" : "string"
                    }
                  },
                  "unicodeLocaleKeys" : {
                    "uniqueItems" : true,
                    "type" : "array",
                    "items" : {
                      "type" : "string"
                    }
                  },
                  "displayLanguage" : {
                    "type" : "string"
                  },
                  "displayScript" : {
                    "type" : "string"
                  },
                  "displayCountry" : {
                    "type" : "string"
                  },
                  "displayVariant" : {
                    "type" : "string"
                  },
                  "extensionKeys" : {
                    "uniqueItems" : true,
                    "type" : "array",
                    "items" : {
                      "type" : "string"
                    }
                  },
                  "iso3Language" : {
                    "type" : "string"
                  },
                  "iso3Country" : {
                    "type" : "string"
                  }
                }
              },
              "ifNoneMatch" : {
                "type" : "array",
                "items" : {
                  "type" : "string"
                }
              },
              "ifUnmodifiedSince" : {
                "type" : "integer",
                "format" : "int64"
              },
              "acceptLanguage" : {
                "type" : "array",
                "items" : {
                  "type" : "object",
                  "properties" : {
                    "range" : {
                      "type" : "string"
                    },
                    "weight" : {
                      "type" : "number",
                      "format" : "double"
                    }
                  }
                }
              },
              "basicAuth" : {
                "type" : "string",
                "writeOnly" : true
              },
              "range" : {
                "type" : "array",
                "items" : {
                  "$ref" : "#/components/schemas/HttpRange"
                }
              },
              "origin" : {
                "type" : "string"
              },
              "accept" : {
                "type" : "array",
                "items" : {
                  "$ref" : "#/components/schemas/MediaType"
                }
              },
              "allow" : {
                "uniqueItems" : true,
                "type" : "array",
                "items" : {
                  "$ref" : "#/components/schemas/HttpMethod"
                }
              },
              "etag" : {
                "type" : "string"
              },
              "expires" : {
                "type" : "integer",
                "format" : "int64"
              },
              "ifMatch" : {
                "type" : "array",
                "items" : {
                  "type" : "string"
                }
              },
              "pragma" : {
                "type" : "string"
              },
              "upgrade" : {
                "type" : "string"
              },
              "vary" : {
                "type" : "array",
                "items" : {
                  "type" : "string"
                }
              },
              "contentType" : {
                "$ref" : "#/components/schemas/MediaType"
              },
              "ifModifiedSince" : {
                "type" : "integer",
                "format" : "int64"
              }
            },
            "additionalProperties" : {
              "type" : "array",
              "items" : {
                "type" : "string"
              }
            }
          },
          "body" : {
            "$ref" : "#/components/schemas/User"
          },
          "statusCode" : {
            "$ref" : "#/components/schemas/HttpStatusCode"
          },
          "statusCodeValue" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      }
    }
  }
}